+++ CHANGELOG.md
@@ -1,10 +1,15 @@
-Recent changes to PracTempo.
+# PracTempo Changelog
 
-## July 2
+## July 6, 2025
+
+* Add floating metronome view.
+* Allow multiple triads to be selected at once (to quickly compare layouts).
+
+## July 2, 2025
 
 * Add floating views for reference.
 * Update README file. 
 
-## May 3
+## May 3, 2025
 
 * Add ability to highlight notes in scales